remove old checkpoint
